Concepts

+++ concepts/src/Cercle.java
@@ -1,5 +1,5 @@
 public class Cercle extends Forme {
-    int rayon;
+    double rayon;
     public double aire(double rayon, double hauteur) {
         return Math.PI * Math.pow(rayon, 2);
     }
